feature : Cars에 CarNames property를 추가하고, 이를 기준으로 생성자를 추가한 후 차들의 기록과 우승자를 제공하는 메소드 추가
